Allow token to be set via settings
The default workflow of the netlify-cli requires you to click a link to allow authentication. Unfortunately you might not always be able to do so (e.g. once the app is deployed onto a server).
In addition, that token is then stored on disk which might require you to click the authentication link more than once if your server has an ephemeral file systems like Heroku.
Luckily the netlify-cli accepts the token to be passed as an argument, removing the need to click the authentication link.